Add ModelTemplate
